Added dataset and Jupyter file
The goal was to best describe the variation in the different types of customers that a wholesale distributor interacts with. We did this by Feature Scaling, optimal number of features using RFECV, KMeans Clustering, PCA, and XGBoost Classifier.

The data set (Wholesale-customers-data.csv)refers to clients of a wholesale distributor. It includes the annual spending in
monetary units on diverse product categories.